loader: have loader_magic be uintptr_t
Twice we've have feedback on this structure not being
correct - folks missed the union statment - so changing
this to be the same size as a pointer to make it more
clear that loaderMagic is the same size as *loaderData.

+++ include/vk_icd.h
@@ -14,7 +14,7 @@
 #define ICD_LOADER_MAGIC   0x01CDC0DE
 
 typedef union _VK_LOADER_DATA {
-  uint32_t loaderMagic;
+  uintptr_t loaderMagic;
   void *loaderData;
 } VK_LOADER_DATA;
